fix: comprehensive bug fixes, wiring, and warning cleanup
Phase 0 — Bug Fixes:
- Add CursorLeft handler to cancel drag/trim when cursor leaves window
- Add mark_sync_dirty() after undo/redo so clips resync to new positions
- Fix 4 NaN panics in partial_cmp().unwrap() (percussion analysis + zoom)
- Wire active_clips count from engine to perf HUD
- Fix get_selected_clip_ids() to resolve region clips for all keyboard
  shortcuts (Delete, Copy, Cut, Duplicate, Split, Nudge, Mute)

Phase 1 — Wire Existing Backends:
- Wire effect param text input commit (parse, clamp, ChangeEffectParamCommand)
- Wire generator param text input commit (ChangeGeneratorParamsCommand)
- Wire effect group rename text input (RenameGroupCommand)
- Wire effect browser paste button (PasteEffects PanelAction + clipboard count)
- Wire effect group/ungroup keyboard shortcuts (Cmd+G / Cmd+Shift+G)

Phase 2 — Dead Code & Warning Cleanup:
- Delete dead code: cached_get_media_length, build_ready_clips_list,
  find_timeline_clip_clone, dead viewport constants, dead get_clip_color
- Fix ~90 compiler warnings across all crates (unused imports, unused
  variables, dead struct fields, clippy lints)
- Suppress deprecated io_surface warnings with migration note
- Suppress dead_code for ported-but-unwired infrastructure

Phase 3 — Port Missing Features:
- Wire MIDI file import via context menu (rfd dialog + MidiFileParser)
- Effect group/ungroup commands fully wired to keyboard shortcuts

+++ crates/manifold-editing/src/service.rs
@@ -274,7 +274,7 @@ impl EditingService {
                     let trimmed = Self::trim_clip_to_region(clip, region, spb);
                     self.clipboard.push(ClipboardEntry {
                         beat_offset: trimmed.start_beat - origin_beat,
-                        layer_offset: trimmed.layer_index - min_layer as i32,
+                        layer_offset: trimmed.layer_index - min_layer,
                         source_clip: trimmed,
                     });
                 }
@@ -939,6 +939,7 @@ impl EditingService {
 
     /// Split clips at region boundaries and return the split commands + interior clip IDs.
     /// Port of C# EditingService.SplitClipsForRegionMove (lines 1135-1143).
+    #[allow(clippy::type_complexity)]
     pub fn split_clips_for_region_move(
         project: &Project,
         region: &SelectionRegion,
